First ArcPy Lab - Geography458 - University of Washington
The first python coding lab for the Geography458 class at the University
of Washington.

The goal of the lab was to take a shape file containing all the census
blocks of Washington State, then go and separate out all the census
blocks according to their county and create a new .shp and .geojson
named accordingly. The second part required me to calculate the total
population of each county and then print out a ranked list in descending
order of the top 10 most populous counties in Washington State